Fix fine-grained permissions config
* disabled by default
* evaluaton context considers fine-grained only when acm is enabled
