+++ hawkbit-repository/hawkbit-repository-jpa/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+org.eclipse.hawkbit.repository.jpa.acm.AccessControllerConfiguration
